fix(invert_sentence)
